Move files around to fix circular dependencies and fix typing issues
