+++ src/main/start-main-application/runnables/setup-lens-proxy.injectable.ts
@@ -11,8 +11,7 @@ import lensProxyPortInjectable from "../../lens-proxy/lens-proxy-port.injectable
 import isWindowsInjectable from "../../../common/vars/is-windows.injectable";
 import showErrorPopupInjectable from "../../electron-app/features/show-error-popup.injectable";
 import { beforeApplicationIsLoadingInjectionToken } from "../runnable-tokens/before-application-is-loading-injection-token";
-import buildVersionInjectable from "../../vars/build-version/build-version.injectable";
-import initializeBuildVersionInjectable from "../../vars/build-version/init.injectable";
+import buildVersionInjectable, { initializeBuildVersionOnMainInjectable } from "../../vars/build-version/build-version.injectable";
 
 const setupLensProxyInjectable = getInjectable({
   id: "setup-lens-proxy",
@@ -70,7 +69,7 @@ const setupLensProxyInjectable = getInjectable({
           return exitApp();
         }
       },
-      runAfter: di.inject(initializeBuildVersionInjectable),
+      runAfter: di.inject(initializeBuildVersionOnMainInjectable),
     };
   },
