+++ src/renderer/stores/init-user-store.injectable.ts
@@ -5,7 +5,7 @@
 import { getInjectable } from "@ogre-tools/injectable";
 import userStoreInjectable from "../../common/user-store/user-store.injectable";
 import { beforeFrameStartsInjectionToken } from "../before-frame-starts/before-frame-starts-injection-token";
-import initializeBuildVersionInjectable from "../vars/build-version/init.injectable";
+import { initializeBuildVersionOnRendererInjectable } from "../vars/build-version.injectable";
 
 const initUserStoreInjectable = getInjectable({
   id: "init-user-store",
@@ -15,7 +15,7 @@ const initUserStoreInjectable = getInjectable({
     return {
       id: "init-user-store",
       run: () => userStore.load(),
-      runAfter: di.inject(initializeBuildVersionInjectable),
+      runAfter: di.inject(initializeBuildVersionOnRendererInjectable),
     };
   },
   injectionToken: beforeFrameStartsInjectionToken,
